// timmyhub-nextjs/src/types/enums.ts
// Enum for User Roles
export enum UserRole {
    CUSTOMER = 'CUSTOMER',
    SELLER = 'SELLER',
    BRAND = 'BRAND',
    SHIPPER = 'SHIPPER',
    ADMIN = 'ADMIN',
    SUPER_ADMIN = 'SUPER_ADMIN',
}

// Enum for Management Tab Types
export enum ManagementTabType {
    LIST = 'list',
    CREATE = 'create',
    UPDATE = 'update',
    DETAIL = 'detail',
}